backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
